// employee/src/media/common/ts/enums/entity-type.enum.ts
export enum MediaEntityTypeEnum {
  LEAVE_REQUEST = 'LEAVE_REQUEST',
  EMPLOYEE_PROFILE = 'EMPLOYEE_PROFILE',
  EMPLOYEE_IDENTIFICATION = 'EMPLOYEE_IDENTIFICATION',
  BULK_IMPORT_INPUT = 'BULK_IMPORT_INPUT',
  BULK_IMPORT_RESULT = 'BULK_IMPORT_RESULT',
  COMPANY_LOGO = 'COMPANY_LOGO',
  EMPLOYEE_INFO_UPDATE = 'EMPLOYEE_INFO_UPDATE',
  RESIGNATION_REQUEST = 'RESIGNATION_REQUEST',
  MISSED_SCAN = 'MISSED_SCAN',
  OVERTIME_REQUEST = 'OVERTIME_REQUEST',
  MISSION_REQUEST = 'MISSION_REQUEST',
  WARNING = 'WARNING',
  DAY_OFF_REQUEST = 'DAY_OFF_REQUEST',
  PAYROLL_BENEFIT_ADJUSTMENT = 'PAYROLL_BENEFIT_ADJUSTMENT',
  MOVEMENT = 'MOVEMENT',
  PAYROLL_DEDUCTION = 'PAYROLL_DEDUCTION',
  BORROW = 'BORROW/PAYBACK_HOUR',
  UPDATE_EMPLOYEE_ATTACHMENT = 'UPDATE_EMPLOYEE_ATTACHMENT'
}